refactor: split owlapy/utils.py into a subpackage
owlapy/utils.py (1989 LOC) was a grab-bag of unrelated concerns
(similarity metrics, expression length, canonical ordering, NNF/CNF/DNF,
syntactic simplification, signature extraction, LRU cache), hurting
navigation and review per IMPROVEMENTS.md 1.2.

Split into owlapy/utils/ with one file per concern (similarity.py,
length.py, ordering.py, nnf.py, simplify.py, signature.py, cache.py).
owlapy/utils/__init__.py re-exports every name previously importable
from the module (verified against the original by exact name-set and
AST-level body diff), so `from owlapy.utils import ...` is unaffected.
No signature changes; full test suite and ruff pass unchanged.
