+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `owl_expression_to_sparql`/`Owl2SparqlConverter` now translate `OWLDataIntersectionOf`, `OWLDataUnionOf`, and `OWLDataComplementOf` data ranges (previously unimplemented, raising `NotImplementedError`) (#242)
 
 ### Changed
+- `owlapy/utils.py` (1989 LOC) is now the package `owlapy/utils/`, split by concern into `similarity.py`, `length.py`, `ordering.py`, `nnf.py`, `simplify.py`, `signature.py`, and `cache.py`. `owlapy.utils`'s `__init__.py` re-exports every name previously importable from the module, so `from owlapy.utils import ...` is unaffected; no signature changes (IMPROVEMENTS.md 1.2)
 - Core library now emits diagnostics through the standard `logging` module instead of `print()`. Progress/status messages log at `INFO`, recoverable issues at `WARNING`, and low-level dumps at `DEBUG`, across `owl_ontology.py`, `owl_reasoner.py`, `render.py`, `utils.py`, and `util_owl_static_funcs.py`. A `NullHandler` is attached to the top-level `owlapy` logger so nothing is printed unless the host application configures logging (opt in via `logging.getLogger("owlapy").setLevel(logging.INFO)`).
 - The `agen_kg` LLM pipeline now also logs through the `logging` module instead of `print()`. The existing `enable_logging` flag keeps its meaning: when set, progress messages are emitted via per-module loggers under `owlapy.agen_kg` and a console handler is attached so output stays visible without extra configuration; `except`-block diagnostics now use `logger.exception(...)` and include tracebacks. Host applications that configure logging themselves can control verbosity via the `owlapy` logger hierarchy.
 - Added `IMPROVEMENTS.md`, a prioritized plan of readability, performance, documentation, and feature improvements for the library
